Changed type of utils_timediff

+++ inc/utils.h
@@ -19,6 +19,6 @@ void utils_mktime(time_t time, char *string);
 void utils_mkresponsetime(time_t time, char *string);
 
 /* Get the difference in seconds between the timespecs; nano-second precision) */
-float utils_timediff(struct timespec end_time, struct timespec start_time);
+double utils_timediff(struct timespec end_time, struct timespec start_time);
 
 #endif
